backend/requirements.txt
Flask==3.0.3
Flask-Reuploaded==1.3.0
opencv_python==4.10.0.84
flask
flask-cors
opencv-python-headless
git+https://github.com/openai/CLIP.git
pillow
torch
torchvision
transformers
tiktok-hashtag-analysis
scikit-learn
joblib
librosa